+++ lightning/src/ln/mod.rs
@@ -28,6 +28,7 @@ pub mod peer_handler;
 pub mod chan_utils;
 pub mod features;
 pub(crate) mod onchaintx;
+pub(crate) mod package;
 
 #[cfg(feature = "fuzztarget")]
 pub mod peer_channel_encryptor;
